refactor: split sound module into separate backend modules
- Extract command-based audio playback to _sound_cmd.py
- Extract sounddevice-based audio playback to _sound_sounddevice.py
- Update main sound.py to orchestrate between backends
- Improve modularity and maintainability
- Preserve all existing functionality and API
